requirements.txt
google-cloud-bigquery
SQLAlchemy 
pandas-gbq
pandas
psycopg2-binary
requests